dj_vercereg/python_env_requirements.txt
Django==1.7
MySQL-python==1.2.5
PyYAML==3.11
Pygments==1.6
django-guardian==1.2.4
django-pdb==0.4.0
django-rest-swagger==0.2.0
django-reversion==1.8.4
django-suit==0.2.11
djangorestframework==2.4.3
requests==2.4.3
six==1.8.0
wsgiref==0.1.2
